More updates for stx-openstack armada manifests
A few updates:
 - Enable yamllint on the application manifests
 - Fix truthy warnings from the linter
 - Fix indentation issue with the horizon conf section and the
   ceilometer db_sync section which invalidates the yaml formatting.
 - Disable the openstack-telemetry chart group as this will not work due
   to dependencies on panko which hasn't been fully integrated yet.

NOTE: Problems highlighted by yamllint in the manifests may not
necessarily be real issues as these are not actually guaranteed to be
standards-compliant YAML files until after they have been run through
the pre-processor (i.e. armada). The goal of using yamllint here is to
catch gross structural errors but any other errors may need to be
evaluated on a case-by-case basis and excluded from the linter after
validation with armada.

Change-Id: Ic604e61e0d6354d56a687c6876cd8d83cae75afc
Story: 2004005
Task: 27801
Signed-off-by: Robert Church <robert.church@windriver.com>