Don't delete nova-ks-endpoint job in nova chart
The nova chart cleans up nova endpoints during nova chart upgrade,
this causes nova-api service to fail to restart. The cleanup for
nova endpoints should only be done in the nova api proxy chart
because they are created by the nova api proxy chart.
This commit updates the armada manifests to remove the deletion
job for nova-ks-endpoints in the nova chart.

Story: 2004520
Task: 28485

Change-Id: Ieea8aaf12eded2e66608012939770505628b6c12
Signed-off-by: Angie Wang <angie.wang@windriver.com>