Support ovs in container
As stx cutovers to containerization, most openstack components run in
containers, but ovs-dpdk running on the host.

This patch is to support ovs running in container, and make it the
default setting. We still support running ovs-dpdk on the host.

For option ovs-dpdk on the host, run follow command before unlock.
```
system modify --vswitch_type ovs-dpdk
```

Change-Id: I111865de17fc82a1e7c269fb97f53a181fb382b2
Story: 2004649
Task: 28610
Co-authored-by: Huifeng Le <huifeng.le@intel.com>
Signed-off-by: chengli3 <cheng1.li@intel.com>