Remove maintenance of two manifest files
manifest-no-tests.yaml and manifest.yaml and identical except for the
"test: enabled" value. We currently only use “false”. The manifest fails
to apply when test enabled is set to true. This issue is being tracked
with https://bugs.launchpad.net/starlingx/+bug/1819021. Once the issue is
resolved with test enabled a system parameter should control setting this
config to true/false.

Story: 2003909
Task: 29878

Change-Id: Iae3450e8285b775350ab05350cbe4cf420e6e4ef
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>