Optimize the installation of ceph-pools-audit
Removes any delays associated with the installation of ceph-pools-audit.
This chart simply installs a CronJob to periodically review and adjust
rbd pool replication based on the desired settings specified in the Ceph
backend.

Here, the dedicated chart group is removed, the chart is specified as
part of the provisioner chart group, the provisioner chart group is now
sequenced with ceph-pools-audit occurring after the rbd-provisioner, and
the ceph-pools-audit waits for the rbd-provisioner app label.

This results in the immediate installation of the chart and we remove
the time delay in waiting for first run of the CronJob which could be up
to five minutes.

Change-Id: Icb8415e4206322df074e0f697a9d800414bd8394
Story: 2004520
Task: 29964
Signed-off-by: Robert Church <robert.church@windriver.com>