Remove magnum from baremetal.
Magnum is no longer packaged on bare metal.

The sysinv and upgrades code related to magnum has been removed.

The helm configuration for magnum remains, although it is not currently
supported in containers either. The magnum-ui is not installed in
platform or containerized horizon so the code to enable it is removed.

Some upgrade code remains, due to the fact that that utility is
in the process of being re-written.

Story: 2004764
Task: 34333
Change-Id: I56873b4e04aac2e7d0cd57909beea00ecc2c1b9a
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>