fix armada stuck at processing osh-openstack-ceph-rgw chart
The issue occur when do swact && lock/unlock standby controller.
After swact && lock/unlock standby controller, the pod created for
job doesn't exist any more. So there is no pod for ceph_rgw.
While armada assumes there is at least 1 pod exist in default, and
will wait for the pod to be up and ready. For chart without pod, we
need explicitly declare the resources that need to be waited for in
armada schema. This declare will override the default wait list in
armada.
Change the timeout value to 300s, which should be enough for the 3
job to be ready.

Closes-Bug: 1833609

Change-Id: I5339406cf914cd54f45b3de5df7ff213e8845bfc
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>