Change rpc_response_max_timeout default to 60 seconds
In stx.1.0, stx-neutron hard-coded the rpc_response_max_timeout value
to 60 seconds. With the migration to containers and upstream neutron,
the default is now set to 600 seconds. To align with the previous
starlingx behavior, the rpc_response_max_timeout set to 60 seconds
by the system through a neutron helm override.

Change-Id: Ibf0f591ac9cb05dac09add37b3c31f6f5b66446d
Closes-Bug: #1836413
Signed-off-by: marvin <weifei.yu@intel.com>