Change cinder bootstrap script
When https is enabled, the cinder helm chart would fail
to apply due to a failure in cinder bootstrap script.
The cinder client which uses the public endpoint to query
the volume types cannot resolve the dns name locally.
This commit changes the cinder template bootstrap script
to use the openstack client instead of the cinder client
to list volume types.

Upstream (openstack-helm) review:
https://review.opendev.org/#/c/671024/

Partial-Bug: 1826583

Change-Id: I63bf1d557c1dd4c93a0063cc9ec1624584079917
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>