Cleanup unused docker image references
The helm charts contain references to images for all
configurations, however some of those configurations
are not being enabled, and so the docker images are never
used.

This change prevents armada from downloading docker images
that are not being used by the armada manifest.

It requires an enhancement in sysinv to handle the null
reference.

The following images are unused and have been replaced
in the manifest with null (or the appropriate reference)

 - kolla/ubuntu-source-nova-novncproxy: referenced by
novnc_assets and must point to the nova image.

 - xrally/xrally-openstack: used when "test" is enabled,
referenced by cinder, ceilometer, glance, heat, keystone,
nova, neutron, panko.

 - openstackhelm/ceph-daemon: referenced by ceph_rgw.

 - openstackhelm/neutron: referenced by openstack-ingress.

 - osixia/keepalived: referenced by openstack-ingress.

 - prom/memcached-exporter: referenced by openstack-memcached.

 - docker: referenced by image_repo_sync image tags in almost
every chart.

 - kbudde/rabbitmq-exporter: referenced by openstack-rabbitmq.

 - prom/mysqld-exporter: referenced by openstack-mariadb

Change-Id: Ide26ddaf3537b8b9595104a683339554aea71b48
Closes-Bug: 1841611
Depends-On: https://review.opendev.org/#/c/680067/
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>