Fix ipv6 address issue causing mariadb ingress not ready
In nginx.tmpl, it not enclose ipv6 addresses in square brackets
resulting in them being unable to be parsed, which cause mariadb
ingress pod could not be ready.

Tested it on both ipv4 and ipv6 simplex setup, it fixes mariadb
ingress not ready issue.

Upstream patch submitted as below
https://review.opendev.org/#/c/710413/

Partial-Bug: 1859641

Change-Id: Ic7726eea671bbedf4f37fbe31965bc8fffd2e8cd
Signed-off-by: Zhipeng Liu <zhipengs.liu@intel.com>