Change stx-openstack services token validation endpoint
If no interface is set, the default is the admin interface.
In a normal deployment both openstack-keystone admin and internal
endpoints are exposed on the same URL.

In case of Distributed Cloud openstack-keystone URL for admin is
overriden to something different than internal URL.

Due to lack of interface, the default `admin` is used, yet the URL is
set to the interal one. The config doesn't point to a valid Keystone
service. Thus Cinder, Nova and Nova-api-proxy cannot validate tokens.

Modify configs for services to point to the internal keystone service.

Closes-Bug: 1875914
Change-Id: I19385f1fe27cf8f20f5ee7b43abf86b220c9d8d6
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>