Optimize nova and neutron helm plugins
The dbapi calls in nova and neutron plugins scale linearly with
the number of worker nodes which results in poor performance on
a large number of nodes system.

Currently, the dbapi calls get invoked for each of worker node.
This commit reduces it to a certain number of calls.

Tested stx-openstack upload and apply on a lab with 6 worker nodes.
The time of override generation for nova reduced from 11s to 0.6s
and for neutron reduced from 24s to 0.2s.

Also tested on vbox and lab has "vf" type of interface. Verified
the content of generated overrides are same as before.

Change-Id: I2d19d30b01e3348d6eb60b8e2681e3a30ef93ebc
Partial-Bug: 1886563
Signed-off-by: Angie Wang <angie.wang@windriver.com>