Update openstack requests to go through ingress
A follow-up commit of https://review.opendev.org/#/c/749624/
to update the requests from nova-api-proxy to nova
via ingress. This requires to enable jobs in nova
to create an additional ingress and ingress-service
for nova-api.

The created ingress for nova-api is called "nova-api-internal"
and nova-api-proxy configuration file is overrided to
listen on nova-api-internal.openstack.svc.cluster.local:80

Change-Id: I2275be8e9458addbf1aedb203a7960f5f8d0b0de
Partial-Bug: 1880777
Signed-off-by: Angie Wang <angie.wang@windriver.com>