Remove kube-system-ingress from openstack operator
When we apply stx-openstack with the 'mode' argument
like `system application apply restore_db`, only
some of the openstack charts must be deployed.
If kube-system-ingress chart groups is specified,
it won't be found in the armada manifest and the
openstack application will always be deployed
in the default way (deploying all the charts),
ignoring the value of the 'mode' argument.

Depends-on: https://review.opendev.org/#/c/698003/
Change-Id: I6791974e337cd3193bf2a75e9d75f48841f0676d
Story: 2006770
Task: 37780
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>