Remove empty package from python-k8sapp-openstack
Packages defined in a spec with no files do not result in an RPM
produced by the build. On a rebuild, the build tools scan the spec and
sees the package defined but does not find a corresponding RPM, and so
flags the package for a rebuild as a result.

This commit removes the empty package definition from the spec.

Partial-Bug: 1910439
Signed-off-by: Don Penney <don.penney@windriver.com>
Change-Id: Ie1f18b1592f8187900624d993434ba04b23cbcff