Address python3 string issues with subprocess
This patch updates our Popen call to enable
newlines for calls that we parse or consume the output for.
Without universal_newlines=True, the output is treated as bytes
under python3 which leads to issues later where we are using it as
strings.

See https://docs.python.org/3/glossary.html#term-universal-newlines

Story: 2006796
Task:  42712

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Ib9f6e9e9bb6ba465993f0ffa4579e86c52716190