Change default Help URL on Horizon
The Help URL at user's dropdown located on the top right corner is
taking the user, by default, to the latest release documentation,
which changes over time and is not the current release used by
stx-openstack. Also this change allows the URL to be easily replaced
as desired via helm-overrides.

Closes-Bug: #1940318
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>
Change-Id: I74b7218356518ac1407091b010949ed90abd169d