Horizon: Fix to show the buttons to Create / Delete / Edit a Role
Openstack is not showing the buttons to Create / Delete / Edit a
Role when using the admin user with admin project, as a result
the admin user can only work with the default Roles created by
Openstack.

Horizon has some features that were rewritten with Angular, the
Role view is one that was rewritten with angular. According to
Horizon documentation, there are necessary configurations that
need to be supplied in Horizon configuration to allow the features
in Angular to work properly.

We found out that the Horizon
REST_API_REQUIRED_SETTINGS  configuration key from
local_settings file is overriden in openstack-armada-app, but it
doesn't have the value OPENSTACK_KEYSTONE_BACKEND
that is necessary to the Role view to work properly.

This is fix has the goal to add the
OPENSTACK_KEYSTONE_BACKEND value in the
REST_API_REQUIRED_SETTINGS configuration key.

Closes-Bug: #1946384
Change-Id: I83563595ee30963ed506685cec24729d9a456268
Signed-off-by: Ricardo Sarto <Ricardo.SoaresSarto@windriver.com>