Create ServiceAccount for pci-irq-affinity-agent
The stx-openstack apply is failing with a timeout condition,
caused by failure while trying to pull the image of the agent from
the local registry. This is caused by missing imagePullSecrets on
the daemonset definition, so this commit creates the ServiceAccount
with the imagePullSecrets and use it on the spec.

Closes-Bug: 1951245
Change-Id: I5e408539e001bbe867812db8c614d47cea1ae225
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>