Pin placement's db_migrate image version
On commit 03d78718966e85eb0094ed48062d9d33f3519f85 of
openstack-helm a new image was introduced to execute the db
migrations as a "latest" form. This commit pins the version of
that image for stx-openstack to a fixed version to avoid that
future image builds of porthole-mysqlclient-utility breaks the
app.

Closes-Bug: #1951444

Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I2b599578858c9212fb61419f166d4db8527f884d