Adapt pci-irq helm chart for oslo_config
The pci-irq-affinity-agent currently uses two mechanisms to
store its configuration parameters, and this is being unified
in [1] to use only oslo_config. As the agent uses a custom
config file in a path not used by default by oslo_config,
it will be needed to pass this path as argument when running the
application, so this commit changes the helm chart to pass the
agent config file with the "--config-file" argument used by
oslo_config.

[1] https://review.opendev.org/c/starlingx/utilities/+/822540

Test Plan:
PASS: Verify that agent definition is created with the
      argument --config-file
PASS: Verify that the agent command script runs without errors
PASS: Verify that the health probe script runs without errors

Regression:
PASS: Verify that OpenStack applies successfully

Story: 2009299
Task: 44214
Change-Id: Ibd9b8c9537ccd46351a4ec511a36c273ae0536b4
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>