Providing overrides for newer images on OSH
After the upversion of OSH, we perceived that stx-openstack is still
downloading the images "openstackhelm/nova" and "openstackhelm/heat".
These are default images on the OSH charts and we want to overwrite them
with stx-heat and stx-nova images so we don't bring unnecessary and
probably unsupported images to the StarlingX registry and stx-openstack
charts.

TEST PLAN
PASS Install stx-openstack and made sure no openstackhelm/* images are
     being downloaded.

Log of the image downloading: https://paste.opendev.org/show/812347/

Closes-Bug: #1958677

Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I365a6f373cced98cb62aa51626a9657abad41b08