Add force_boot command to rabbit start template
Currently, if a multi-node cluster is shut down unexpectedly,
RabbitMQ is not able to boot and sync with the other nodes.

The purpose of this change is to add the patches to enable
the possibility to use the rabbitmqctl force_boot command
to recover RabbitMQ from an unexpected cluster shut down.

Test plan:
PASS: Shutdown and start a multi-node RabbitMQ cluster

Regression:
PASS: OpenStack can be applied successfully
PASS: RabbitMQ nodes can join the RabbitMQ cluster

Story: 2009784
Task: 44290

Ref:
[0] https://www.rabbitmq.com/rabbitmqctl.8.html#force_boot

Signed-off-by: Maik Catrinque
<maik.wandercatrinqueandrade@windriver.com>
Change-Id: Ib9dc89483fc14510f96ce56559cc31bd7b28d6ac