Add method to check if OpenStack is HTTPS ready
This change adds a _is_openstack_https_ready function to the
openstack.py plugin. It verifies if the platform has https_enabled set
to true and checks if three certificates are present: openstack,
openstack_ca and ssl_ca. If both conditions are met, it returns True.

Test Plan
PASS: Build OpenStack and verify that the generated tarball contains the
added code.
PASS: Apply the built tarball.

Signed-off-by: Gustavo Santos <gustavofaganello.santos@windriver.com>
Change-Id: I28e4bdb0785ae453830a426a731f14a0b80a0d47