Removing pci-irq-affinity lifecycle hooks
This code is being removed as part of an effort to move the
pci-irq-affinity-agent into a openstack application container, instead
of a platform service.

Since now the service will no longer run on the platform but on a
container instead, there is no need to update the platform service
config file after the application apply and remove actions.

TEST PLAN:
PASS: Successfully build the stx-openstack application
PASS: Successfully apply the stx-openstack application

Story: 2009299
Task: 44569

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Ic711a2e9d7c32e6e217de936b5b1c141f85e1128