Update RabbitMQ probes
The healthcheck used to liveness and readiness probes is considered
intrusive and can produce false positives[0]. It's just returning that
the pod is healthy if it's fully booted and whether the pod is waiting
for a peer to appear to sync tables, it cannot succeed and eventually
enters into a pod restart deadlock.
The command is also deprecated and will be removed in future version.
Updating the probes based on current recommenation from community[1].
Because of that, we change the liveness to check the rabbit connection
to check if the rabbit is connected and running and also changed the
readiness to ping the pod after boot to see if it's ready.

Ref:
[0] https://www.rabbitmq.com/monitoring.html#deprecations
[1] https://www.rabbitmq.com/monitoring.html#health-checks

Test plan:
PASS: Close rabbitmq connection inside pod and verify if Liveness fail
and rebooted the pod.
PASS: Readiness check pod state and verify if can ping it to check if
the pod is health after get up.
PASS: Build on Debian

Story: 2009892
Task: 44674

Signed-off-by: Arthur Luz de Avila <arthur.luzdeavila@windriver.com>
Change-Id: Ib98df7e1135b7b5cee50d63a9ab06eb4a961cd4b