Decrease terminationGracePeriodSeconds on glance-api
The glance-api pod has a terminationGracePeriodSeconds
of 600s(10min) and the others services has 30s. This high
terminationGracePeriodSeconds may cause timeout in some
cases and there is no reason for this high
terminationGracePeriodSeconds.
The terminationGracePeriodSeconds has been introduced on
https://review.opendev.org/c/openstack/openstack-helm/+/469974
but there is no explanation why it is too high.
the Glance don't gracefully terminate due to a authenticate
with kube-apiserver  with the error: "Unable to authenticate
the request" err="[invalid bearer token, Token has been invalidated]"
Looking after it I saw that it can be a timing error between the
communication of the kube-apiserver and the glance-api when is in
terminating state. Normally, the glance api terminate gracefully
before 30sec, but if the error of Unable to authenticate appear, it
kill faster to keep the things working like the other services does.

Test Plan:
PASS: Build on centOS
PASS: Build on Debian
PASS: Apply chart

Story: 2009959
Task: 44927

Signed-off-by: Arthur Luz de Avila <arthur.luzdeavila@windriver.com>
Change-Id: Id4ef301c9ee1fce9f78ee09b3ce7f896d59a5a96