Making stx_admin (osh) helm v3 compatible
The patch added by this change will make the stx_admin job
syntax compatible with FluxCD (helmv3). It is necessary to add
a hook-weight annotation to the resource to guarantee that all
required resources will be up and running by the time of the
job creation.

Test Plan:
PASS - build-helm-charts.sh builds a basic FluxCD and a complete Armada
       app tarball
PASS - application upload and overrides generated
PASS - application apply and check if the new hooks takes place
PASS - application remove/delete

Closes-Bug: 1990188

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I83137f54eb18ba01b880e093a77b987283e5d907