OSH Patch: Fixing cinder release hooks weights
The relation of dependency for cinder release resources is not working
with helmv3 since several jobs have post-install hooks and are
dependencies of other jobs and deployments that have no hooks.

The jobs/deployments without hooks are deployed during an installation
phase that is never complete since the dependency jobs are hooked to be
deployed on post-install phase.

This change includes helm-hooks for the boostrap job and the api,
scheduler and volume deployments. The weights will define the order each
one will be deployed.

Test Plan:
PASS - build-helm-charts.sh builds a basic FluxCD and a complete Armada
       app tarball
PASS - application upload and overrides generated
PASS - application apply/remove/delete

Closes-Bug: 1990305

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I146a5608c60082239d85d1cd8f0f2545c0c62091