OSH Patch: Fixing placement release hooks
The relation of dependency for placement resources is not working
with helmv3 since for a couple of jobs the helm-hooks and weights are
not implemented correctly.

This is already implemented on openstack-helm main branch, so this patch
is simply bringing those updates to openstack-armada-app.

Test Plan:
PASS - build-helm-charts.sh builds a basic FluxCD app tarball
PASS - application upload and overrides generated
PASS - application apply/remove/delete

Closes-Bug: 1990398

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Id95ca55f254319bb5bcd6712817b2258126820ed