Decode cmd return string for keyring password
When running the application on Debian, now on Python 3 strings are
stored as Byte Strings instead of Unicode Strings. The return of cmd
execution that is stored as a Byte String is problematic for keyring
password saving function, that requires an Unicode string.

This change simply decode the bytes to Unicode, so the plugin is now
compatible with Python 3. By Duck Typing it, the plugin will still be
compatible with Python 2 for ensuring upgrades are still possible.

TEST PLAN:
PASS - Build stx-openstack-helm-fluxcd package
PASS - Run build-helm-charts.sh (build app)
PASS - Upload stx-openstack app on Debian
PASS - Upload stx-openstack app on CentOS

Closes-Bug: 1998231

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Ib28bca757c1946513e0b7bc57d4c22937842587a