Copy ceph-etc through ceph-etc-pools-audit
In latest versions of the platform-integ-apps app the
current provisioner has been changed to ceph-csi [1].
In this provisioner the ceph-etc ConfigMap no longer
exists. In order to create the ceph-etc in the openstack
namespace we now use the ceph-etc-pools-audit ConfigMap.
A storyboard has been created [2] in order to keep track
of the adaptation that needs to take place to make use of
the correct ceph resources.

Test Plan:
PASS - Build stx-openstack tarball.
PASS - Upload stx-openstack.
PASS - Apply stx-openstack and check that the ceph-etc
       ConfigMap has been successfully created.

Closes-Bug: 1998230

[1] - https://review.opendev.org/c/starlingx/platform-armada-app/+/858737
[2] - https://storyboard.openstack.org/#!/story/2010456

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I52a5ded7fcc14a1aafd6eee0c92e1db29dba52e1