Remove explicit set of docker0 as tunnel interface
Last month [1] was merged removing the default docker network bridge to
avoid IP conflicts with addresses already in use. Since StarlingX no
longer runs containers with docker it is a reasonable change.

Later it was noticed that neutron-openvswitch-agent-init.sh access a
network.interface.tunnel that is explicitly set to docker0 since the
original Armada manifests [2] of stx-openstack application. This was
inherited from Armada to FluxCD during the application migration.

Since this default network bridge is no longer available, this static
override needs to be updated. We are relying now on the fallback
mechanism to search for this interface, as peer [3] and [4].
In case there is any future network use case that requires a different
approach, it could also be possible to use the application plugins to
retrieve the cluster-host interface and dynamically override this field.

[1] https://review.opendev.org/c/starlingx/config-files/+/865329
[2] https://opendev.org/starlingx/openstack-armada-app/commit/b7d0b3ed0c9e6ceac86d63088ac783dd0adecf7b
[3] https://github.com/openstack/openstack-helm/blob/master/neutron/values.yaml#L110
[4] https://github.com/openstack/openstack-helm/blob/master/neutron/templates/bin/_neutron-openvswitch-agent-init.sh.tpl#L423

TEST PLAN:
PASS - Build stx-openstack fluxcd helm charts
PASS - Upload and Apply stx-openstack
PASS - Check that neutron-ovs-agent-controller pod spawns successfully

Closes-Bug: 1999426

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I169d8408420483bbf4e6c59c2cf70be5da039481