Add patch to libvirt setup script
On Debian this libvirt and qemu users/groups setup changed and it
seems to be easier to maintain now, so we can drop a libvirt script
setup step.

Previously, on CentOS, there was no mismatch between the container kvm
GID and the host kvm GID because both had our qemu rpm installed and
it was forcing the GID to 36 [1]. On Debian it was removed at all [2],
relying now on the debian/systemd patch [3][4] that already creates
the group and sets the device (/dev/kvm) permissions and ownership
accordingly.

Since on Debian sytemd is already creating the kvm group and handling
the /dev/kvm permission and ownership, we can remove this step from our
libvirt container setup script and rely on Debian installation defaults.

[1] https://opendev.org/starlingx/integ/src/branch/master/virt/qemu/centos/qemu-kvm.spec#L722
[2] https://salsa.debian.org/qemu-team/qemu/-/commit/cb8737ef48a37eddf12ac199b46f9034273ba6d3
[3] https://salsa.debian.org/systemd-team/systemd/-/commit/4fc3fa53bfa6e16ceb6cd312f49003839b56144a
[4] https://bugs.launchpad.net/ubuntu/+source/gnome-boxes/+bug/1767302/comments/18

Test Plan:
PASS - Build openstack-helm-infra
PASS - Build stx-openstack-fluxcd package
PASS - Build stx-openstack helm charts
PASS - Upload/Apply/Remove the application
PASS - Check that the script skipped the kvm device permission set
PASS - Check that the host kvm device has the correct permissions and
       ownership.
PASS - Check the container and host users and groups

Partial-Bug: 1999445

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I47e5be5f34989f932902d2b7f97ef23bedac3260