Update tox.ini to work with tox 4
This change will allow this repo to pass zuul now
that this has merged:
https://review.opendev.org/c/zuul/zuul-jobs/+/866943

Tox 4 deprecated whitelist_externals.
Replace whitelist_externals with allowlist_externals

Removed the py27 zuul job target.

Updated the base tox.ini so that the showconfig
will pass for the pep8 target in the tox.ini in the subdir.

Partial-Bug: #2000399

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I4227eae6dc9e796f784984d31f575297ae0f2838