Creating folder structure for STX-O upversion
In preparation for the STX-Openstack upversion to ANTELOPE, there's a
need to decouple the Openstack packages from the Platform ones, so that
the development can occur without impacting the current version. This
decoupling allows the platform and applications to leverage different
versions of OpenStack releases for their respective packages.

On this upversion process, we can't upversion one package at a time,
since that would certainly break the build, testing and usage of
StarlingX. We need to have a way in which the development of STX-O can
occur without impacting the rest of the platform.

To achieve this, a new folder named "openstack" has been created
under the openstack-armada-app repo, containing all packages from the
upstream repository, under the openstack folder [1].
This ensures that the ongoing upversion process does not disrupt the
platform's functionalities and prevents any potential breakages during
the transition.

It should be noted that this change only serves as a placeholder for
future upversions and does not impact the current packages, as they're
not going to be delivered to the ISO, or the STX-O Helm charts, until
the upversion is completed.

This commit lays the groundwork for future work on the STX-O upversion,
promoting flexibility and modularization within the StarlingX packages.

The SRC_DIR and the BASE_SRCREV were updated for each package in order
for the build to work. The BASE_SRCREV was set to the latest merged
change by the time that this review was created [2]

After the work on the upversion and the STX-O clients containerization
is done, the 48138 task (under this same story) can be worked on to
completely remove the STX-O clients from the platform.

[1] https://opendev.org/starlingx/upstream/src/branch/master/openstack
[2] https://review.opendev.org/c/starlingx/openstack-armada-app/+/885301

Test Plan:
PASS: Make sure that the packages are not found if they're commented
      on the debian_pkg_dirs file
PASS: Make sure that the images are not found if they're commented on
      the debian_stable_docker_images.inc
PASS: Make sure packages can be build:
      - openstack-pkg-tools
      - rabbitmq-server
      - python-wsme
      - openstack-ras
      - python-openstacksdk
      - python-oslo-messaging
      - python-osc-lib
      - barbican
      - keystone
      - horizon
      - python-aodhclient
      - python-barbicanclient
      - python-cinderclient
      - python-glanceclient
      - python-gnocchiclient
      - python-heatclient
      - python-ironicclient
      - python-keystoneclient
      - python-neutronclient
      - python-novaclient
      - python-openstackclient
      - python-pankoclient
PASS: Make sure images can be build:
      - stx-aodh
      - stx-ironic
      - stx-barbican
      - stx-ceilometer
      - stx-cinder
      - stx-glance
      - stx-gnocchi
      - stx-heat
      - stx-horizon
      - stx-keystone
      - stx-neutron
      - stx-nova
      - stx-openstackclients
      - stx-placement
      - stx-platformclients

Story: 2010774
Task: 48115

Change-Id: I077a814382fb21bd4b36fac7c20ee041718433f3
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>