Upversion python-openstackclient to v6.2.0-1
As part of the STX-Openstack upversion to ANTELOPE, the python-
openstackclient is being upversioned to version 6.2.0-1, which is the
latest supported on ANTELOPE [1].

All patches were updated so the package can be build. Note that on the
0002-Add-location-parameter-for-volume-backup-creation.patch, the end
part of the patch had to be removed [2], because on this new version,
the openstackclient it added a `**kwargs` argument [3], making it
unnecessary to specify the location argument.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-openstackclient
[2] https://opendev.org/starlingx/openstack-armada-app/src/branch/master/upstream/openstack/python-openstackclient/debian/patches/0002-Add-location-parameter-for-volume-backup-creation.patch#L119
[3] https://salsa.debian.org/openstack-team/clients/python-openstackclient/-/blob/debian/6.2.0-1/openstackclient/volume/v2/volume_backup.py#L171

Test Plan:
PASS: Build the python-openstackclient package

Story: 2010715
Task: 48137

Change-Id: I3318b6b54acbcfce67a3f56edd7076fb1b5ac2a2
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>